Refactor component imports and restructure type definitions
