--- src/app/interface.ts
@@ -1,17 +1,0 @@
-export interface simplifiedProduct {
-  _id: string;
-  name: string;
-  imageUrl: string;
-  price: number;
-  slug: string;
-  categoryName: string;
-}
-export interface fullProduct {
-  _id: string;
-  imageUrl: string;
-  price: number;
-  slug: string;
-  categoryName: string;
-  name: string;
-  description: string;
-}
