+++ src/components/ImageGallery.tsx
