+++ types/index.d.ts
@@ -1,0 +1,18 @@
+interface simplifiedProduct {
+    _id: string;
+    name: string;
+    imageUrl: string;
+    price: number;
+    slug: string;
+    categoryName: string;
+}
+
+interface fullProduct {
+    _id: string;
+    imageUrl: string;
+    price: number;
+    slug: string;
+    categoryName: string;
+    name: string;
+    description: string;
+}
